* Fix Sagorne Creststrider <Shaman Trainer> "Train" gossip. * Fix Trial of the Sea Lion - chest location + add bubble object. * Fix Quest Mage Summoner (missing SAI for creature). * Allow flying for Pridewing consort. * Delete wrong Mirthil spawns in durotar cave. * Fix Deadmines door lever not opening doors. * Fix quest Let them eat crows - add plaguehound pet script. * Fix quest Agamaggan's Agility not giving a buff reward. * Fix quest Agamaggan's Strength not giving a buff reward. * Fix quest Spirit of the Wind not giving a buff reward. * Fix quest Wisdom of Agamaggan not giving a buff reward. * Fix quest Rising Spirit not giving a buff reward. * Fix quest Razorhide not giving a buff reward. * Fix quest Investigate the Alchemist Shop - script weren't working. * Fix quest Investigate the Blue Recluse - script weren't working. * Fix quest Westfall Stew - NPC quest item displaying. * Fix quest Kobold Candles - NPC quest item displaying. * Fix quest Gold Dust Exchange - NPC quest item displaying. * Fix quest Goretusk Liver Pie - NPC quest item displaying. * Fix quest The Woodland Protector - NPC quest item displaying. * Disable quest Relic of the Earthen Ring - 3.3.x content. * Fix Watch Commander Zalaphil - too big DMG. * Fix movemap issue with invisible walls. * Fix dungeon graveyards teleporting to wrong zones. * Fix zone display in some dungeons. * Fix guardians of NPC not defending when master is attacked. * Fix Bleeding Hollow Tormentor not despawning his worg. * Fix quest The Wrap Rifts not despawning after 3 summons. * Fix quest Blueleaf Tubers - item interaction + script. * Fix quest A Spirit Guide - NPC speed. * Fix zone exploration. * Make creatures fall after death.

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -838,7 +838,7 @@ void ObjectMgr::CheckCreatureTemplate(CreatureTemplate const* cInfo)
     if (!displayScaleEntry)
         sLog->outErrorDb("Creature (Entry: %u) does not have any existing display id in Modelid1/Modelid2/Modelid3/Modelid4.", cInfo->Entry);
 
-    for (int k = 0; k < MAX_KILL_CREDIT; ++k)
+    for (uint8 k = 0; k < MAX_KILL_CREDIT; ++k)
     {
         if (cInfo->KillCredit[k])
         {
